refactor: split guide content into individual files
