--- content/guideContent.js
@@ -1,140 +1,0 @@
-export const pageContent = {
-    'how-to-grow': {
-        title: "How to Grow Dragon Fruit: The Complete Start-to-Fruit Guide",
-        description: "Learn everything you need to know to grow dragon fruit at home, from choosing a variety and planting cuttings to pruning, pollination, and harvesting.",
-        author: "Growing Dragon Fruit",
-        lastUpdated: "June 28, 2025",
-        crumbs: [{ label: "How to Grow", path: "how-to-grow" }],
-        toc: ["Choosing a Variety", "Planting: Cuttings vs. Seeds", "Ideal Climate & Location", "Pots vs. In-Ground", "Watering Wisely", "Pruning for Production", "Pollination & Harvesting", "Overwintering in Cold Climates", "Common Beginner Mistakes"],
-        content: [
-          { heading: "Choosing the Right Dragon Fruit Variety", paragraphs: ["Your dragon fruit journey begins with selecting the right variety, a crucial first step that will influence flavor, maintenance, and your overall success. There are hundreds of cultivars, but they generally fall into three categories based on skin and flesh color: Hylocereus undatus (pink skin with white flesh), Hylocereus costaricensis (pink skin with red or pink flesh, also known as Pitaya Roja), and Hylocereus megalanthus (yellow skin with white flesh, often called Pitaya Amarilla).", "For beginners, the most important factor is choosing a <strong>self-pollinating</strong> (or self-fertile) variety. This means you only need one plant to produce fruit. Excellent self-pollinating choices include 'American Beauty', 'Zamorano', 'Sugar Dragon', and 'Halley's Comet'. These varieties are known for being vigorous growers and reliable producers, making them perfect for your first foray into dragon fruit cultivation. Varieties that require cross-pollination will need a genetically different dragon fruit plant nearby to set fruit, which is a consideration for those with limited space. Our Varieties guide provides a deep dive into the specific flavors and growth habits of each."] },
-          { heading: "Planting: The Cutting vs. Seed Debate", paragraphs: ["There are two ways to start a dragon fruit plant: from a seed or from a cutting. While you can technically grow a plant from the tiny black seeds in a store-bought fruit, be prepared for a long wait. A seed-grown plant can take anywhere from five to seven years to mature enough to produce fruit, and the resulting fruit may not be true to its parent.","For this reason, <strong>growing from cuttings is the universally recommended method.</strong> A cutting is a segment of a mature, healthy plant that will grow into a genetic clone of its parent. This means you know exactly what kind of fruit you'll get, and more importantly, a cutting can start producing fruit in as little as one to three years. To plant a cutting, ensure it's at least 6-12 inches long. Let the cut end dry in a shady, dry place for about a week. This 'callusing' process is vital to prevent rot. Once callused, plant the cutting about 2-3 inches deep in your chosen soil mix."], image: "https://placehold.co/800x600/A7F3D0/14532D?text=Healthy+Cutting+(800x600)" },
-          { heading: "Ideal Climate, Sunlight, and Location", paragraphs: ["Dragon fruit is a subtropical cactus that thrives in warmth and light. It is best suited for USDA hardiness zones 9-11 where it can be grown outdoors year-round. The absolute most important requirement is sunlight—at least <strong>6 to 8 hours of direct sunlight per day</strong> is necessary for the plant to photosynthesize effectively and produce fruit. An east-facing location that gets plenty of morning sun is ideal. If you live in an extremely hot climate (e.g., desert regions of Arizona or California), providing some afternoon shade with a 30% shade cloth can be beneficial to prevent the stems from getting scorched by the intense late-day sun."] },
-          { heading: "Growing in Pots vs. In the Ground", paragraphs: [], table: { headers: ["Method", "Best For", "Pros", "Cons"], rows: [
-                        ["Container Growing", "Colder climates (Zones below 9); Limited space.", "Portability for winter protection; Full control over soil.", "Dries out faster; Requires more frequent watering/fertilizing."],
-                        ["In-Ground Growing", "Warmer climates (Zones 9-11); Ample space.", "Allows for a larger root system; More stable temperature.", "Requires well-draining native soil; Less protection from pests/frost."]
-                    ]}},
-          { heading: "Watering Wisely: The Secret to a Healthy Cactus", paragraphs: ["Here's the most common mistake new growers make: treating dragon fruit like a desert cactus. While it is a cactus, it's a tropical one, and it requires more water than its desert cousins. However, it is still highly susceptible to root rot from overwatering. The key is balance.", "The golden rule is to <strong>water deeply when the top 2-3 inches of soil are completely dry.</strong> During the active growing season (spring and summer), you may need to water every few days, especially if it's in a pot in a hot climate. In the fall and winter, when the plant is dormant, reduce watering significantly—perhaps only once every few weeks. Always check the soil first! Sticking your finger deep into the soil is the most reliable method. The plant would rather be a little too dry than a little too wet."] },
-          { heading: "Pruning for Abundant Fruit Production", paragraphs: ["An unpruned dragon fruit plant will become a tangled, unproductive mess. Pruning is absolutely essential to direct the plant's energy towards flowering and fruiting. Your goal is to establish a main upright stem and then encourage an 'umbrella' of arching, fruit-bearing branches at the top of your trellis. As the main stem grows, remove any small side shoots to focus energy on vertical growth. Once the stem reaches the top of the trellis, snip the tip. This 'apical cut' will trigger the plant to send out multiple new branches from that point. These will become your primary fruiting arms. Each year, after the main fruiting season, perform a 'clean-up' prune. Remove any dead, damaged, or overcrowded stems to improve air circulation and make way for new, productive growth."] },
-          { heading: "The Magic of Pollination and the Joy of Harvest", paragraphs: ["Many dragon fruit flowers are nocturnal, opening at dusk and closing by mid-morning. It's a spectacular sight. If your variety is self-pollinating, you're all set. If it requires cross-pollination, you'll need to play the role of a bee. Using a small, soft brush, gently collect the yellow pollen from the anthers of a flower on one plant and transfer it to the stigma (the central, often greenish part) of a flower on a different variety. The video below provides a great visual guide.", "After successful pollination, a small fruitlet will begin to form. The fruit is typically ready to harvest <strong>30 to 50 days after flowering.</strong> The key indicator of ripeness is color. The fruit should be fully and brightly colored (e.g., a deep, uniform pink for a pink variety). The small leafy 'wings' on the fruit will also start to wither slightly. To harvest, give the fruit a gentle twist. If it's ripe, it should detach easily. If you have to pull hard, it needs more time."], video: { id: 'v_zWk_2mfnY', title: "How to Hand-Pollinate Dragon Fruit" } },
-          { heading: "Overwintering Dragon Fruit in Colder Climates", paragraphs: ["If you live in a zone colder than 9, you must protect your plant from frost. Dragon fruit can tolerate a brief dip to just above freezing, but a hard frost will kill it. For container-grown plants, the process is straightforward. Once nighttime temperatures consistently drop below 50°F (10°C), it's time to bring your plant indoors. Place it near the sunniest window in your house, preferably a south-facing one. Reduce watering to a bare minimum—just enough to keep the soil from becoming bone dry (perhaps once a month). Do not fertilize during this dormant period. Your plant may look a little sad over the winter, but it's simply resting. In the spring, once the danger of frost has passed, you can gradually reintroduce it to the outdoors."] },
-          { heading: "Common Beginner Mistakes to Avoid", paragraphs: ["We've all been there! Here are the top three pitfalls to avoid on your journey:", "<blockquote><strong>1. Overwatering:</strong> This is the number one plant killer. Remember, let the topsoil dry out completely. Better to err on the side of too dry than too wet.</blockquote>", "<blockquote><strong>2. Insufficient Sunlight:</strong> Six hours of direct sun is the minimum for fruit. If your plant is all stem and no flower, it's likely craving more light.</blockquote>", "<blockquote><strong>3. Inadequate Support:</strong> A flimsy trellis will collapse under the weight of a mature plant. Build a strong, permanent structure from the very beginning. Your future self will thank you.</blockquote>"] },
-        ],
-        blogLinks: [
-            {title: "5 Signs You're Overwatering Your Dragon Fruit", description: "Learn to recognize the subtle signs before it's too late.", path: "blog/5-signs-of-overwatering"},
-            {title: "A Step-by-Step Guide to Your First Pruning", description: "That first cut is the most important. Learn how to create the perfect shape for your plant.", path: "blog/pruning-for-shape"}
-        ],
-        cta: { heading: "Ready to Grow?", subheading: "Grab our free care calendar to stay on track!", buttonText: "Get Your Free Care Calendar" }
-    },
-    'soil-guide': {
-        title: "Best Soil for Dragon Fruit: The Ultimate DIY Mix Guide",
-        description: "Discover the perfect DIY soil mix for your dragon fruit. Learn about drainage, pH, and the best store-bought options for containers and in-ground planting.",
-        author: "Growing Dragon Fruit",
-        lastUpdated: "June 28, 2025",
-        crumbs: [{ label: "Soil Guide", path: "soil-guide" }],
-        toc: ["Why Soil is Critical", "The Three Pillars of Perfect Soil", "DIY Soil Mix Recipes", "Choosing Store-Bought Mixes", "Testing and Adjusting Soil pH", "Soil for Pots vs. In-Ground", "Amending Soil Seasonally"],
-        content: [
-          { heading: "Why Getting the Soil Right is Critical", paragraphs: ["To understand dragon fruit soil, you must understand the plant's nature. It is an <strong>epiphytic cactus</strong>, which means in its native jungle environment, it often grows on other trees, using its roots to anchor itself in bark crevices and pockets of decaying organic matter. This gives us our most important clue: it is not adapted to dense, wet, traditional garden soil. It hates having 'wet feet'. The right soil mix is the foundation of a healthy plant; it provides stability, prevents fatal root rot, allows roots to breathe, and delivers nutrients efficiently."] },
-          { heading: "The Three Pillars of Perfect Pitaya Soil", paragraphs: ["Every successful dragon fruit soil mix is built on three pillars: drainage, pH, and nutrients.", "<strong>1. Drainage:</strong> This is the most critical factor. The soil must be loose, gritty, and allow water to pass through freely. Without excellent drainage, water will stagnate around the roots, leading to rot. Key components for drainage include sand, perlite, pumice, and orchid bark.", "<strong>2. pH Level:</strong> Dragon fruit prefers a slightly acidic to neutral soil, with a target pH range between <strong>6.0 and 7.2</strong>. Most standard potting mixes and composts fall within this range, but it's wise to test your soil if you are having issues with nutrient uptake.", "<strong>3. Organic Nutrition:</strong> While it needs drainage, it's not a desert cactus. It thrives with a good amount of organic matter. Decomposed compost, worm castings, and aged manure provide a slow, steady release of the essential nutrients the plant needs for vigorous growth."] },
-          { heading: "Our Favorite DIY Soil Mix Recipes", paragraphs: ["Creating your own mix is easy, cost-effective, and gives you complete control. Here are two proven recipes:", "<strong>The Classic Mix (Balanced):</strong> \n- 40% Coarse Sand or Perlite (for drainage) \n- 40% Aged Compost or Worm Castings (for nutrients) \n- 20% Potting Soil or Coco Coir (for structure and moisture retention)", "<strong>The Pro-Potting Mix (For Containers):</strong> \n- 50% Orchid Bark Mix (contains bark, perlite, and charcoal) \n- 30% High-Quality Cactus/Succulent Mix \n- 20% Aged Compost", "As you can see in the image below, the ideal texture is loose and gritty, not dense and muddy."], image: "https://placehold.co/800x600/E9D5FF/3730A3?text=Gritty+Soil+Mix+(800x600)" },
-          { heading: "Choosing the Best Store-Bought Mix", paragraphs: ["If you prefer the convenience of a bagged mix, look for high-quality <strong>cactus and succulent soil</strong>. Read the ingredients list and avoid mixes that are very heavy on peat moss, as it can compact and retain too much moisture over time. Even with the best store-bought mix, we strongly recommend improving it. A simple and effective upgrade is to mix in one extra part of perlite or pumice for every two parts of the bagged soil. This small step can dramatically improve drainage and prevent future problems."] },
-          { heading: "Testing and Adjusting Your Soil's pH", paragraphs: ["If your plant is struggling with yellowing leaves despite proper fertilization, the issue could be pH. When soil pH is too high (alkaline) or too low (acidic), it 'locks up' nutrients, making them unavailable to the plant. You can buy a simple soil pH testing kit from any garden center. To lower pH (make it more acidic), you can amend the soil with elemental sulfur or use acid-loving plant fertilizers. To raise pH (make it more alkaline), you can add garden lime or wood ash. Always make small adjustments and re-test after a few weeks, as changing soil pH is a gradual process."] },
-          { heading: "Soil Considerations for Pots vs. In-Ground", paragraphs: ["For potted plants, drainage is paramount. Your container <strong>must</strong> have multiple large drainage holes. Do not add a layer of rocks at the bottom; this is a myth and actually hinders drainage. Use a light, airy soil mix like the 'Pro-Potting Mix' above. When planting in the ground, assess your native soil. If you have heavy clay, you must amend it heavily or, better yet, create a dedicated raised bed or mound for your dragon fruit. Dig a large hole (at least 2x2 feet) and backfill it entirely with your custom, well-draining dragon fruit soil mix."] },
-          { heading: "Amending and Revitalizing Your Soil Seasonally", paragraphs: ["The nutrients in a pot of soil are finite. Each year, it's a good practice to revitalize the soil, especially for hungry plants like dragon fruit. In the spring, you can gently top-dress your containers with a 1-2 inch layer of fresh compost or worm castings. This will provide a slow-release source of nutrients for the upcoming growing season. For in-ground plants, mulching with organic material serves a similar purpose, breaking down over time to enrich the soil."] },
-        ],
-        blogLinks: [],
-        cta: { heading: "Get Your Soil Right!", subheading: "Now that you have the foundation, learn how to feed your plant.", buttonText: "Learn About Fertilizing" }
-    },
-    'fertilizer': {
-        title: "Fertilizing Dragon Fruit for Growth & Bloom: A Complete Guide",
-        description: "Learn the best fertilizing schedule for dragon fruit. Understand NPK ratios, organic vs. synthetic options, and how to read your plant for deficiencies.",
-        author: "Growing Dragon Fruit",
-        lastUpdated: "June 28, 2025",
-        crumbs: [{ label: "Fertilizer", path: "fertilizer" }],
-        toc: ["Understanding NPK by Growth Stage", "Organic vs. Synthetic Fertilizers", "A Year-Round Fertilizing Schedule", "Micronutrients: The Unsung Heroes", "Reading Your Plant for Deficiencies"],
-        content: [
-            { heading: "Decoding NPK: Feeding for Growth vs. Fruit", paragraphs: ["Fertilizer isn't just 'plant food'; it's a strategic tool to tell your plant what to do. The three numbers on a fertilizer bag, N-P-K, represent Nitrogen (N), Phosphorus (P), and Potassium (K). Each plays a different role. <strong>Nitrogen (N)</strong> fuels green, leafy growth. <strong>Phosphorus (P)</strong> is crucial for root development and flowering. <strong>Potassium (K)</strong> supports overall plant function and fruit quality.", "During the plant's <strong>growth phase</strong> (when it's young or in early spring), use a balanced fertilizer like a 16-16-16 or a 20-20-20 to encourage strong, thick, green stems. Once the plant is mature and you want to encourage <strong>fruiting</strong>, switch to a low-nitrogen, high-phosphorus 'bloom booster' formula, like a 2-10-10 or 0-10-10. Giving a mature plant too much nitrogen will result in a beautiful, lush green plant with zero fruit."] },
-            { heading: "Organic vs. Synthetic: Which is Better?", paragraphs: ["This is a common debate, and the answer is: both have their place. Many expert growers use a hybrid approach. They build their soil with rich organic compost and then use a targeted synthetic bloom booster during the summer to get the best results."], table: { headers: ["Fertilizer Type", "Pros", "Cons"], rows: [
-                        ["Organic", "Improves soil health; Slow-release reduces risk of burning.", "Slower to act; Can be less precise."],
-                        ["Synthetic", "Provides a quick, targeted boost; Precise nutrient ratios.", "Can burn plants if overused; Can build up salts in the soil."]
-                    ]}, image: "https://placehold.co/800x600/FEF9C3/713F12?text=Compost+vs+Fertilizer+(800x600)" },
-            { heading: "A Sample Year-Round Fertilizing Schedule", paragraphs: ["This schedule is a general guideline; adjust based on your specific climate and plant's growth."], table: {
-                    headers: ["Season", "Action Required"],
-                    rows: [
-                        ["Early Spring", "Apply balanced slow-release granular fertilizer and top-dress with compost."],
-                        ["Late Spring to Mid-Summer", "Fertilize every 2-4 weeks with a balanced liquid fertilizer or switch to a high-phosphorus bloom booster."],
-                        ["Late Summer / Early Fall", "Give one last small feeding of a balanced fertilizer to help the plant recover."],
-                        ["Late Fall / Winter", "<strong>Stop fertilizing completely.</strong> The plant is resting and doesn't need the extra nutrients."]
-                    ]
-                }},
-            { heading: "Micronutrients: The Unsung Heroes of Plant Health", paragraphs: ["While NPK gets all the attention, micronutrients are just as vital for a healthy plant. <strong>Calcium</strong> is crucial for building strong cell walls and preventing issues like stem rot. <strong>Magnesium</strong> is a key component of chlorophyll and is essential for photosynthesis. <strong>Iron</strong> is also necessary for chlorophyll production, and a deficiency often results in yellowing new growth.", "The best way to provide these is through rich, organic compost and by using a liquid fertilizer that contains micronutrients. A monthly foliar spray (spraying the leaves) with a diluted liquid seaweed or kelp fertilizer is an excellent way to provide a direct boost of these essential trace elements."] },
-            { heading: "Reading Your Plant: Signs of Nutrient Deficiency", paragraphs: ["Your plant will tell you what it needs. Learning to read the signs is a key skill. \n\n- <strong>Yellowing, especially on older stems:</strong> This is often a classic sign of <strong>Nitrogen deficiency</strong>. The plant is moving nitrogen from old growth to new growth. \n- <strong>Stunted growth and poor flowering:</strong> Could indicate a <strong>Phosphorus deficiency</strong>. \n- <strong>Yellowing around the edges of stems:</strong> Can be a sign of <strong>Potassium deficiency</strong>. \n- <strong>Yellowing on new growth (while old growth stays green):</strong> This is a classic sign of an <strong>Iron deficiency</strong>. \n- <strong>Thin, weak, pale new growth:</strong> This is a general sign of overall under-fertilization. The plant simply doesn't have the fuel to produce strong new stems. If you see these signs, it's time to adjust your feeding regimen."] },
-        ],
-        blogLinks: [
-            {title: "Can I Use Tomato Fertilizer on My Dragon Fruit?", description: "When is it okay and when is it a bad idea? We break down the NPK numbers.", path: "blog/can-i-use-tomato-fertilizer"}
-        ],
-        cta: { heading: "Feed Your Way to a Bountiful Harvest!", subheading: "Keep your plant care on schedule with our free calendar.", buttonText: "Download Free Calendar" }
-    },
-    'trellis-guide': {
-        title: "Dragon Fruit Trellis Designs: A DIY & Training Guide",
-        description: "Explore the best trellis designs for dragon fruit, including a DIY plan for the popular post and wheel system. Learn how to train your climbing cactus for maximum fruit production.",
-        author: "Growing Dragon Fruit",
-        lastUpdated: "June 28, 2025",
-        crumbs: [{ label: "Trellis Guide", path: "trellis-guide" }],
-        toc: ["Why a Trellis is Non-Negotiable", "Comparing Trellis Designs", "DIY Plan: The Post and Wheel Trellis", "Training Your Plant Up the Trellis", "Common Trellising Mistakes", "Alternative Trellis Ideas"],
-        content: [
-            { heading: "Why a Trellis is a Non-Negotiable Investment", paragraphs: ["A common beginner mistake is underestimating the size and weight of a mature dragon fruit plant. This is not a plant you can simply grow in a hanging basket. Dragon fruit is a vigorous climbing cactus that produces aerial roots to cling to surfaces, scaling its way towards sunlight. Without a strong support system, the plant will become a tangled, sprawling mess on the ground. This leads to a host of problems: the stems are more susceptible to rot and pests, air circulation is poor, and fruit production will be minimal or nonexistent.", "A proper trellis accomplishes several critical goals: it provides the necessary vertical support, it improves air circulation to prevent fungal diseases, and it allows the fruit-bearing branches to cascade downwards in an 'umbrella' shape. This cascading form is the ideal structure for maximizing sunlight exposure and making harvesting easy. Investing in a strong trellis from day one is investing in the long-term health and productivity of your plant."] , image: "https://placehold.co/800x600/DBEAFE/1E3A8A?text=Trellis+with+Branches+(800x600)"},
-            { heading: "Comparing Common Trellis Designs", paragraphs: ["While you can get creative, most successful trellis designs fall into a few categories:"], table: { headers: ["Design", "Pros", "Cons"], rows: [
-                        ["Post and Wheel", "Excellent air circulation; Easy to harvest from all sides.", "Requires a strong, deeply set post."],
-                        ["T-Bar", "Simple to construct from wood.", "Can get crowded if not pruned well."],
-                        ["Wall-Mounted / Fence", "Saves space; Can look very ornamental.", "Difficult to prune the back; Reduced air circulation."]
-                    ]}},
-            { heading: "DIY Plan: How to Build a Post and Wheel Trellis", paragraphs: ["This is a fantastic weekend project that will last for years. The video below gives a great overview of the process. \n\n<strong>Materials & Tools:</strong> \n- One 10-12 foot long, 6x6 inch pressure-treated wood or concrete post. \n- One bag of concrete mix. \n- A shovel and post-hole digger. \n- A level. \n- A sturdy top support (e.g., a repurposed bicycle rim with spokes removed, a welded rebar square, or a custom-built wooden cross). \n- Heavy-duty bolts or lag screws to attach the top support. \n\n<strong>Steps:</strong> \n1. **Dig the Hole:** Dig a hole at least 2-3 feet deep. The deeper, the better for stability. \n2. **Set the Post:** Place the post in the center of the hole. Use your level to ensure it is perfectly vertical. \n3. **Pour Concrete:** Mix the concrete according to the package directions and pour it into the hole, filling it completely. Brace the post so it remains level while the concrete cures (at least 24-48 hours). \n4. **Attach the Top:** Once the concrete is fully cured, securely bolt your chosen top support to the post. Your trellis is now ready for planting!"], video: { id: 's2j5h5Q6_pA', title: 'Building a DIY Dragon Fruit Trellis' } },
-            { heading: "Training Your Plant Up the Trellis", paragraphs: ["Patience is key during the training phase. After planting one to four cuttings at the base of your new trellis, select the strongest, most vigorous stem from each cutting to be the main 'trunk'. As this main stem grows, remove any other side shoots that sprout from its lower sections. This focuses all the plant's energy on reaching the top of the trellis. Use soft, flexible ties (like cloth strips, gardening velcro, or pantyhose) to loosely secure the stem to the post every foot or so. Do not tie it too tightly, as the stem will thicken over time. Once the main stem has grown about a foot higher than the top support, prune off the growing tip. This will stimulate the plant to branch out, creating the cascading umbrella of fruiting arms."] },
-            { heading: "Common Trellising Mistakes to Avoid", paragraphs: ["<strong>1. Building it Too Short:</strong> A trellis that is only 4-5 feet tall is not enough. Aim for at least 6-7 feet of height above the ground. \n<strong>2. Using Flimsy Materials:</strong> A mature dragon fruit plant laden with fruit is incredibly heavy. A thin wooden stake or a standard tomato cage will buckle and fail. Overbuild it from the start. \n<strong>3. Tying Stems Too Tightly:</strong> This can constrict and damage the stem as it grows, potentially killing the plant. The ties are just for guidance, not for strangulation."] },
-            { heading: "Alternative and Unconventional Trellis Ideas", paragraphs: ["Don't have space for a massive post? You can get creative. For smaller spaces, a very sturdy cattle panel or wire mesh arched between two posts can create a 'tunnel' trellis. Some growers successfully use the pillars of a patio or pergola as the main support, attaching a top structure to them. The key principles remain the same: provide a strong vertical support and a stable top platform for the branches to cascade over."] },
-        ],
-        blogLinks: [
-            {title: "A Step-by-Step Guide to Your First Pruning", description: "That first cut is the most important. Learn how to create the perfect shape for your plant.", path: "blog/pruning-for-shape"}
-        ],
-        cta: { heading: "Build the Perfect Support System!", subheading: "A good trellis is a long-term investment in your plant's health.", buttonText: "Explore Dragon Fruit Varieties" }
-    },
-    'varieties': {
-        title: "Dragon Fruit Varieties & Reviews: From American Beauty to Yellow Dragon",
-        description: "Explore popular dragon fruit varieties like American Beauty, Sugar Dragon, and Yellow Dragon. Learn about their flavor, pollination needs, and which is right for you.",
-        author: "Growing Dragon Fruit",
-        lastUpdated: "June 28, 2025",
-        crumbs: [{ label: "Varieties", path: "varieties" }],
-        toc: ["Pollination: The Most Important Factor", "American Beauty", "Sugar Dragon (S8)", "Yellow Dragon (Palora)", "Physical Graffiti", "Zamorano", "Where to Buy Authentic Cuttings"],
-        content: [
-            { heading: "Pollination: The First Thing to Consider", paragraphs: ["Before you fall in love with a variety's flavor or color, you must understand its pollination needs. This is the single most important factor for a home grower. Dragon fruit varieties are either <strong>self-fertile</strong> (can pollinate themselves and produce fruit with just one plant) or <strong>self-sterile</strong> (require pollen from a different dragon fruit variety to produce fruit).", "If you have limited space and only plan on growing one plant, <strong>you must choose a self-fertile variety.</strong> If you have space for multiple plants, you can explore the world of self-sterile varieties, which often have unique and complex flavors. However, you will need to plant at least two different varieties and be prepared to hand-pollinate them. In the guide below, we clearly state the pollination needs for each popular variety."] },
-            { heading: "American Beauty (Hylocereus guatemalensis)", paragraphs: ["<strong>Pollination:</strong> Self-Fertile \n\nAmerican Beauty is one of the most popular and reliable varieties for home growers, and for good reason. It's a vigorous, fast-growing plant that produces stunning magenta-fleshed fruit. The flavor is a step up from the common store-bought white-fleshed varieties, offering a refreshing sweetness with a hint of berry. It's known for being a consistent and heavy producer once established. Because it's self-fertile, it's a perfect choice for beginners who want a guaranteed harvest without the complexity of cross-pollination. Its bright pink skin and deep purple flesh make it a visual showstopper both on the vine and on the plate."], image: 'https://placehold.co/800x600/FBCFE8/831843?text=American+Beauty+(800x600)'},
-            { heading: "Sugar Dragon (S8)", paragraphs: ["<strong>Pollination:</strong> Self-Fertile \n\nAs the name implies, Sugar Dragon is renowned for its exceptional sweetness. Many enthusiasts consider it one of the best-tasting dragon fruits available. The fruit is typically smaller than American Beauty, but it packs an intense, rich, grape-like flavor with a high Brix (sugar content) level. The flesh is a deep, dark red. The plant itself is a moderately vigorous grower and is self-fertile, making it another excellent choice for those with a single plant. If your top priority is flavor intensity over fruit size, Sugar Dragon is an unbeatable variety."], image: 'https://placehold.co/800x600/F472B6/831843?text=Sugar+Dragon+(800x600)'},
-            { heading: "Yellow Dragon (Palora) (Hylocereus megalanthus)", paragraphs: ["<strong>Pollination:</strong> Self-Fertile \n\nThe 'queen' of sweet dragon fruits. The Yellow Dragon variety from Palora, Ecuador, is famous for its bright yellow skin and translucent white flesh. It is consistently rated as one of the sweetest and most delicious of all pitayas, with a flavor reminiscent of lychee, pear, and kiwi. The fruit contains larger, crunchier seeds than its pink-skinned cousins. While the plant's stems are thinner and more sensitive to intense sun than other varieties, its self-fertility and incredible flavor make it a highly sought-after plant for any serious hobbyist."], image: 'https://placehold.co/800x600/FEF08A/854D0E?text=Yellow+Dragon+(800x600)'},
-            { heading: "Physical Graffiti", paragraphs: ["<strong>Pollination:</strong> Self-Sterile (Requires a Pollinator) \n\nPhysical Graffiti is a perfect example of a variety that's worth the extra effort of cross-pollination. It produces beautiful fruit with a pink skin and a stunning purple-fuchsia flesh. The flavor is complex and excellent, often described as having berry and grape notes. To get fruit from Physical Graffiti, you must have another variety nearby to use as a pollen donor (American Beauty or Zamorano are good partners). You will need to hand-pollinate, transferring pollen from your donor plant to the Physical Graffiti flower. For gardeners with multiple plants, this variety is a must-have for its unique color and taste."], image: 'https://placehold.co/800x600/DDD6FE/4C1D95?text=Physical+Graffiti+(800x600)'},
-            { heading: "Zamorano", paragraphs: ["<strong>Pollination:</strong> Self-Fertile \n\nZamorano is a workhorse variety that is prized for its reliability and large, flavorful fruit. It produces impressive red-fleshed fruit that can often exceed one pound. The flavor is sweet and well-balanced. As a self-fertile variety, it's a low-maintenance choice for beginners, but it's also an excellent pollen donor for any self-sterile varieties you may want to grow, like Physical Graffiti. It's a vigorous grower and a great all-around choice for consistent production and great taste."] },
-            { heading: "Where to Buy Authentic Cuttings", paragraphs: ["The source of your cuttings matters. Buying from a reputable online nursery or a local grower who specializes in dragon fruit is crucial. Be wary of unidentified cuttings on auction sites, as you may end up with a low-quality, non-fruiting variety. Look for sellers who clearly label their varieties and can provide information about their pollination needs. A healthy cutting should be firm, green, and free of any soft spots or blemishes. Investing in a quality cutting from a good source is the best way to start your dragon fruit journey on the right foot."] },
-        ],
-        blogLinks: [],
-        cta: { heading: "Found Your Perfect Pitaya?", subheading: "Now learn exactly how to care for it with our complete guides.", buttonText: "Get Your Free Care Calendar" }
-    },
-    'pests': {
-        title: "Dragon Fruit Pests & Diseases: Identification & Control Guide",
-        description: "Identify and control common dragon fruit pests and diseases. Learn how to prevent root rot, stem canker, and deal with pests like mealybugs and ants.",
-        author: "Growing Dragon Fruit",
-        lastUpdated: "June 28, 2025",
-        crumbs: [{ label: "Pests & Diseases", path: "pests" }],
-        toc: ["Root Rot: The #1 Enemy", "Stem Rot & Sunburn", "Identifying Common Pests", "Natural Prevention Methods", "Dealing with Larger Pests"],
-        content: [
-            { heading: "Root Rot: The Silent Killer and How to Prevent It", paragraphs: ["The most common and devastating issue for dragon fruit is not a pest, but a fungal disease: root rot, caused by overwatering. Because it happens below the soil, the first signs you see are often advanced: yellowing, wilting stems that are soft and mushy at the base. By then, it can be too late. <strong>Prevention is the only cure.</strong> This comes down to two things: using an extremely well-draining soil mix (as detailed in our Soil Guide) and disciplined watering. Always allow the top 2-3 inches of soil to dry out completely before watering again. If you do suspect root rot, you must act fast. Unpot the plant, cut away any black, mushy roots, and repot in fresh, dry soil. Often, the best bet is to take healthy cuttings from the unaffected upper stems and start over."], image: 'https://placehold.co/800x600/FDE68A/713F12?text=Yellowing+Stems+(800x600)'},
-            { heading: "Common Stem Issues: Rot, Canker, and Sunburn", paragraphs: ["<strong>Stem Rot/Canker:</strong> You might notice soft, discolored spots on the stems, sometimes with a reddish-brown ooze. This is typically a bacterial or fungal infection. The best treatment is immediate surgery. Using a sterile knife, cut out the entire affected area, scooping it out until you only see healthy, green tissue. Dusting the wound with cinnamon, a natural fungicide, can help it dry and heal. \n\n<strong>Sunburn:</strong> This appears as a bleached, yellowish or whitish patch on stems that receive the most intense sun. While minor sunburn isn't fatal, it can stress the plant and create a weak spot for infections to enter. In extremely hot climates, providing some afternoon shade with a 30% shade cloth can be very beneficial. A healthy plant with good nutrition is also less susceptible to sunburn."] },
-            { heading: "Identifying and Controlling Common Pests", paragraphs: ["Dragon fruit is relatively tough, but a few pests can become a nuisance. \n\n- <strong>Ants:</strong> If you see ants marching up your plant, they are almost always a secondary issue. They are 'farming' other pests, like aphids or mealybugs, for their sweet secretions (honeydew). Get rid of the primary pest, and the ants will leave. \n\n- <strong>Mealybugs & Aphids:</strong> These small, sap-sucking insects often hide in the joints between stems. Look for white, cottony masses (mealybugs) or tiny green or black dots (aphids). For a minor infestation, a cotton swab dipped in rubbing alcohol will kill them on contact. For larger outbreaks, a spray of insecticidal soap or neem oil is effective. \n\n- <strong>Thrips and Mites:</strong> These are tiny pests that are hard to see, but you'll see their damage: scarred, scaly, or 'rusty' looking patches on the fruit and new growth. Regular preventative sprays with neem oil during the growing season is the best way to keep them at bay."], image: "https://placehold.co/800x600/FECACA/7F1D1D?text=Mealybugs+on+Stem+(800x600)" },
-            { heading: "The Best Defense: Natural and Preventative Methods", paragraphs: ["A healthy plant is a pest-resistant plant. The best way to deal with pests and diseases is to create an environment where they can't thrive. \n\n- <strong>Improve Air Circulation:</strong> Prune your plant annually to prevent a dense, tangled mess. Good airflow helps stems dry out quickly, making it harder for fungal spores to take hold. \n\n- <strong>Water at the Base:</strong> Avoid overhead watering. Splashing water from the soil onto the stems can transfer fungal spores directly to your plant. Water the soil, not the plant. \n\n- <strong>Regular Inspections:</strong> Spend a few minutes each week closely inspecting your plant, especially in the nooks and crannies. Catching a small pest problem early is much easier than dealing with a full-blown infestation. \n\n- <strong>Neem Oil is Your Friend:</strong> A regular, preventative spray of neem oil (every 2-4 weeks during the growing season) acts as a natural fungicide and insecticide, creating a first line of defense against most common issues."] },
-            { heading: "Dealing with Larger Pests (Birds and Rodents)", paragraphs: ["As your fruit begins to ripen, you may attract some larger, uninvited guests. Birds, squirrels, and even rats can be drawn to the sweet, colorful fruit. The most effective way to protect your harvest is with physical barriers. You can purchase 'organza bags' or 'fruit protection bags' online. These are small mesh bags that you can place over each individual fruit as it develops. This allows the fruit to ripen while keeping pests out. For larger-scale protection, you can drape bird netting over your entire trellis structure, securing it at the bottom so pests can't crawl under."] },
-        ],
-        blogLinks: [
-             {title: "5 Signs You're Overwatering Your Dragon Fruit", description: "Learn to recognize the subtle signs before it's too late.", path: "blog/5-signs-of-overwatering"},
-        ],
-        cta: { heading: "Keep Your Plants Healthy and Thriving!", subheading: "Our free calendar helps you track preventative care tasks.", buttonText: "Download Your Free Calendar" }
-    },
-};
