+++ content/guides/fertilizer.js
@@ -1,0 +1,30 @@
+export default {
+        title: "Fertilizing Dragon Fruit for Growth & Bloom: A Complete Guide",
+        description: "Learn the best fertilizing schedule for dragon fruit. Understand NPK ratios, organic vs. synthetic options, and how to read your plant for deficiencies.",
+        author: "Growing Dragon Fruit",
+        lastUpdated: "June 28, 2025",
+        crumbs: [{ label: "Fertilizer", path: "fertilizer" }],
+        toc: ["Understanding NPK by Growth Stage", "Organic vs. Synthetic Fertilizers", "A Year-Round Fertilizing Schedule", "Micronutrients: The Unsung Heroes", "Reading Your Plant for Deficiencies"],
+        content: [
+            { heading: "Decoding NPK: Feeding for Growth vs. Fruit", paragraphs: ["Fertilizer isn't just 'plant food'; it's a strategic tool to tell your plant what to do. The three numbers on a fertilizer bag, N-P-K, represent Nitrogen (N), Phosphorus (P), and Potassium (K). Each plays a different role. <strong>Nitrogen (N)</strong> fuels green, leafy growth. <strong>Phosphorus (P)</strong> is crucial for root development and flowering. <strong>Potassium (K)</strong> supports overall plant function and fruit quality.", "During the plant's <strong>growth phase</strong> (when it's young or in early spring), use a balanced fertilizer like a 16-16-16 or a 20-20-20 to encourage strong, thick, green stems. Once the plant is mature and you want to encourage <strong>fruiting</strong>, switch to a low-nitrogen, high-phosphorus 'bloom booster' formula, like a 2-10-10 or 0-10-10. Giving a mature plant too much nitrogen will result in a beautiful, lush green plant with zero fruit."] },
+            { heading: "Organic vs. Synthetic: Which is Better?", paragraphs: ["This is a common debate, and the answer is: both have their place. Many expert growers use a hybrid approach. They build their soil with rich organic compost and then use a targeted synthetic bloom booster during the summer to get the best results."], table: { headers: ["Fertilizer Type", "Pros", "Cons"], rows: [
+                        ["Organic", "Improves soil health; Slow-release reduces risk of burning.", "Slower to act; Can be less precise."],
+                        ["Synthetic", "Provides a quick, targeted boost; Precise nutrient ratios.", "Can burn plants if overused; Can build up salts in the soil."]
+                    ]}, image: "https://placehold.co/800x600/FEF9C3/713F12?text=Compost+vs+Fertilizer+(800x600)" },
+            { heading: "A Sample Year-Round Fertilizing Schedule", paragraphs: ["This schedule is a general guideline; adjust based on your specific climate and plant's growth."], table: {
+                    headers: ["Season", "Action Required"],
+                    rows: [
+                        ["Early Spring", "Apply balanced slow-release granular fertilizer and top-dress with compost."],
+                        ["Late Spring to Mid-Summer", "Fertilize every 2-4 weeks with a balanced liquid fertilizer or switch to a high-phosphorus bloom booster."],
+                        ["Late Summer / Early Fall", "Give one last small feeding of a balanced fertilizer to help the plant recover."],
+                        ["Late Fall / Winter", "<strong>Stop fertilizing completely.</strong> The plant is resting and doesn't need the extra nutrients."]
+                    ]
+                }},
+            { heading: "Micronutrients: The Unsung Heroes of Plant Health", paragraphs: ["While NPK gets all the attention, micronutrients are just as vital for a healthy plant. <strong>Calcium</strong> is crucial for building strong cell walls and preventing issues like stem rot. <strong>Magnesium</strong> is a key component of chlorophyll and is essential for photosynthesis. <strong>Iron</strong> is also necessary for chlorophyll production, and a deficiency often results in yellowing new growth.", "The best way to provide these is through rich, organic compost and by using a liquid fertilizer that contains micronutrients. A monthly foliar spray (spraying the leaves) with a diluted liquid seaweed or kelp fertilizer is an excellent way to provide a direct boost of these essential trace elements."] },
+            { heading: "Reading Your Plant: Signs of Nutrient Deficiency", paragraphs: ["Your plant will tell you what it needs. Learning to read the signs is a key skill. \n\n- <strong>Yellowing, especially on older stems:</strong> This is often a classic sign of <strong>Nitrogen deficiency</strong>. The plant is moving nitrogen from old growth to new growth. \n- <strong>Stunted growth and poor flowering:</strong> Could indicate a <strong>Phosphorus deficiency</strong>. \n- <strong>Yellowing around the edges of stems:</strong> Can be a sign of <strong>Potassium deficiency</strong>. \n- <strong>Yellowing on new growth (while old growth stays green):</strong> This is a classic sign of an <strong>Iron deficiency</strong>. \n- <strong>Thin, weak, pale new growth:</strong> This is a general sign of overall under-fertilization. The plant simply doesn't have the fuel to produce strong new stems. If you see these signs, it's time to adjust your feeding regimen."] },
+        ],
+        blogLinks: [
+            {title: "Can I Use Tomato Fertilizer on My Dragon Fruit?", description: "When is it okay and when is it a bad idea? We break down the NPK numbers.", path: "blog/can-i-use-tomato-fertilizer"}
+        ],
+        cta: { heading: "Feed Your Way to a Bountiful Harvest!", subheading: "Keep your plant care on schedule with our free calendar.", buttonText: "Download Free Calendar" }
+};
