+++ content/guides/how-to-grow.js
@@ -1,0 +1,27 @@
+export default {
+        title: "How to Grow Dragon Fruit: The Complete Start-to-Fruit Guide",
+        description: "Learn everything you need to know to grow dragon fruit at home, from choosing a variety and planting cuttings to pruning, pollination, and harvesting.",
+        author: "Growing Dragon Fruit",
+        lastUpdated: "June 28, 2025",
+        crumbs: [{ label: "How to Grow", path: "how-to-grow" }],
+        toc: ["Choosing a Variety", "Planting: Cuttings vs. Seeds", "Ideal Climate & Location", "Pots vs. In-Ground", "Watering Wisely", "Pruning for Production", "Pollination & Harvesting", "Overwintering in Cold Climates", "Common Beginner Mistakes"],
+        content: [
+          { heading: "Choosing the Right Dragon Fruit Variety", paragraphs: ["Your dragon fruit journey begins with selecting the right variety, a crucial first step that will influence flavor, maintenance, and your overall success. There are hundreds of cultivars, but they generally fall into three categories based on skin and flesh color: Hylocereus undatus (pink skin with white flesh), Hylocereus costaricensis (pink skin with red or pink flesh, also known as Pitaya Roja), and Hylocereus megalanthus (yellow skin with white flesh, often called Pitaya Amarilla).", "For beginners, the most important factor is choosing a <strong>self-pollinating</strong> (or self-fertile) variety. This means you only need one plant to produce fruit. Excellent self-pollinating choices include 'American Beauty', 'Zamorano', 'Sugar Dragon', and 'Halley's Comet'. These varieties are known for being vigorous growers and reliable producers, making them perfect for your first foray into dragon fruit cultivation. Varieties that require cross-pollination will need a genetically different dragon fruit plant nearby to set fruit, which is a consideration for those with limited space. Our Varieties guide provides a deep dive into the specific flavors and growth habits of each."] },
+          { heading: "Planting: The Cutting vs. Seed Debate", paragraphs: ["There are two ways to start a dragon fruit plant: from a seed or from a cutting. While you can technically grow a plant from the tiny black seeds in a store-bought fruit, be prepared for a long wait. A seed-grown plant can take anywhere from five to seven years to mature enough to produce fruit, and the resulting fruit may not be true to its parent.","For this reason, <strong>growing from cuttings is the universally recommended method.</strong> A cutting is a segment of a mature, healthy plant that will grow into a genetic clone of its parent. This means you know exactly what kind of fruit you'll get, and more importantly, a cutting can start producing fruit in as little as one to three years. To plant a cutting, ensure it's at least 6-12 inches long. Let the cut end dry in a shady, dry place for about a week. This 'callusing' process is vital to prevent rot. Once callused, plant the cutting about 2-3 inches deep in your chosen soil mix."], image: "https://placehold.co/800x600/A7F3D0/14532D?text=Healthy+Cutting+(800x600)" },
+          { heading: "Ideal Climate, Sunlight, and Location", paragraphs: ["Dragon fruit is a subtropical cactus that thrives in warmth and light. It is best suited for USDA hardiness zones 9-11 where it can be grown outdoors year-round. The absolute most important requirement is sunlight—at least <strong>6 to 8 hours of direct sunlight per day</strong> is necessary for the plant to photosynthesize effectively and produce fruit. An east-facing location that gets plenty of morning sun is ideal. If you live in an extremely hot climate (e.g., desert regions of Arizona or California), providing some afternoon shade with a 30% shade cloth can be beneficial to prevent the stems from getting scorched by the intense late-day sun."] },
+          { heading: "Growing in Pots vs. In the Ground", paragraphs: [], table: { headers: ["Method", "Best For", "Pros", "Cons"], rows: [
+                        ["Container Growing", "Colder climates (Zones below 9); Limited space.", "Portability for winter protection; Full control over soil.", "Dries out faster; Requires more frequent watering/fertilizing."],
+                        ["In-Ground Growing", "Warmer climates (Zones 9-11); Ample space.", "Allows for a larger root system; More stable temperature.", "Requires well-draining native soil; Less protection from pests/frost."]
+                    ]}},
+          { heading: "Watering Wisely: The Secret to a Healthy Cactus", paragraphs: ["Here's the most common mistake new growers make: treating dragon fruit like a desert cactus. While it is a cactus, it's a tropical one, and it requires more water than its desert cousins. However, it is still highly susceptible to root rot from overwatering. The key is balance.", "The golden rule is to <strong>water deeply when the top 2-3 inches of soil are completely dry.</strong> During the active growing season (spring and summer), you may need to water every few days, especially if it's in a pot in a hot climate. In the fall and winter, when the plant is dormant, reduce watering significantly—perhaps only once every few weeks. Always check the soil first! Sticking your finger deep into the soil is the most reliable method. The plant would rather be a little too dry than a little too wet."] },
+          { heading: "Pruning for Abundant Fruit Production", paragraphs: ["An unpruned dragon fruit plant will become a tangled, unproductive mess. Pruning is absolutely essential to direct the plant's energy towards flowering and fruiting. Your goal is to establish a main upright stem and then encourage an 'umbrella' of arching, fruit-bearing branches at the top of your trellis. As the main stem grows, remove any small side shoots to focus energy on vertical growth. Once the stem reaches the top of the trellis, snip the tip. This 'apical cut' will trigger the plant to send out multiple new branches from that point. These will become your primary fruiting arms. Each year, after the main fruiting season, perform a 'clean-up' prune. Remove any dead, damaged, or overcrowded stems to improve air circulation and make way for new, productive growth."] },
+          { heading: "The Magic of Pollination and the Joy of Harvest", paragraphs: ["Many dragon fruit flowers are nocturnal, opening at dusk and closing by mid-morning. It's a spectacular sight. If your variety is self-pollinating, you're all set. If it requires cross-pollination, you'll need to play the role of a bee. Using a small, soft brush, gently collect the yellow pollen from the anthers of a flower on one plant and transfer it to the stigma (the central, often greenish part) of a flower on a different variety. The video below provides a great visual guide.", "After successful pollination, a small fruitlet will begin to form. The fruit is typically ready to harvest <strong>30 to 50 days after flowering.</strong> The key indicator of ripeness is color. The fruit should be fully and brightly colored (e.g., a deep, uniform pink for a pink variety). The small leafy 'wings' on the fruit will also start to wither slightly. To harvest, give the fruit a gentle twist. If it's ripe, it should detach easily. If you have to pull hard, it needs more time."], video: { id: 'v_zWk_2mfnY', title: "How to Hand-Pollinate Dragon Fruit" } },
+          { heading: "Overwintering Dragon Fruit in Colder Climates", paragraphs: ["If you live in a zone colder than 9, you must protect your plant from frost. Dragon fruit can tolerate a brief dip to just above freezing, but a hard frost will kill it. For container-grown plants, the process is straightforward. Once nighttime temperatures consistently drop below 50°F (10°C), it's time to bring your plant indoors. Place it near the sunniest window in your house, preferably a south-facing one. Reduce watering to a bare minimum—just enough to keep the soil from becoming bone dry (perhaps once a month). Do not fertilize during this dormant period. Your plant may look a little sad over the winter, but it's simply resting. In the spring, once the danger of frost has passed, you can gradually reintroduce it to the outdoors."] },
+          { heading: "Common Beginner Mistakes to Avoid", paragraphs: ["We've all been there! Here are the top three pitfalls to avoid on your journey:", "<blockquote><strong>1. Overwatering:</strong> This is the number one plant killer. Remember, let the topsoil dry out completely. Better to err on the side of too dry than too wet.</blockquote>", "<blockquote><strong>2. Insufficient Sunlight:</strong> Six hours of direct sun is the minimum for fruit. If your plant is all stem and no flower, it's likely craving more light.</blockquote>", "<blockquote><strong>3. Inadequate Support:</strong> A flimsy trellis will collapse under the weight of a mature plant. Build a strong, permanent structure from the very beginning. Your future self will thank you.</blockquote>"] },
+        ],
+        blogLinks: [
+            {title: "5 Signs You're Overwatering Your Dragon Fruit", description: "Learn to recognize the subtle signs before it's too late.", path: "blog/5-signs-of-overwatering"},
+            {title: "A Step-by-Step Guide to Your First Pruning", description: "That first cut is the most important. Learn how to create the perfect shape for your plant.", path: "blog/pruning-for-shape"}
+        ],
+        cta: { heading: "Ready to Grow?", subheading: "Grab our free care calendar to stay on track!", buttonText: "Get Your Free Care Calendar" }
+};
