+++ content/guides/pests.js
@@ -1,0 +1,19 @@
+export default {
+        title: "Dragon Fruit Pests & Diseases: Identification & Control Guide",
+        description: "Identify and control common dragon fruit pests and diseases. Learn how to prevent root rot, stem canker, and deal with pests like mealybugs and ants.",
+        author: "Growing Dragon Fruit",
+        lastUpdated: "June 28, 2025",
+        crumbs: [{ label: "Pests & Diseases", path: "pests" }],
+        toc: ["Root Rot: The #1 Enemy", "Stem Rot & Sunburn", "Identifying Common Pests", "Natural Prevention Methods", "Dealing with Larger Pests"],
+        content: [
+            { heading: "Root Rot: The Silent Killer and How to Prevent It", paragraphs: ["The most common and devastating issue for dragon fruit is not a pest, but a fungal disease: root rot, caused by overwatering. Because it happens below the soil, the first signs you see are often advanced: yellowing, wilting stems that are soft and mushy at the base. By then, it can be too late. <strong>Prevention is the only cure.</strong> This comes down to two things: using an extremely well-draining soil mix (as detailed in our Soil Guide) and disciplined watering. Always allow the top 2-3 inches of soil to dry out completely before watering again. If you do suspect root rot, you must act fast. Unpot the plant, cut away any black, mushy roots, and repot in fresh, dry soil. Often, the best bet is to take healthy cuttings from the unaffected upper stems and start over."], image: 'https://placehold.co/800x600/FDE68A/713F12?text=Yellowing+Stems+(800x600)'},
+            { heading: "Common Stem Issues: Rot, Canker, and Sunburn", paragraphs: ["<strong>Stem Rot/Canker:</strong> You might notice soft, discolored spots on the stems, sometimes with a reddish-brown ooze. This is typically a bacterial or fungal infection. The best treatment is immediate surgery. Using a sterile knife, cut out the entire affected area, scooping it out until you only see healthy, green tissue. Dusting the wound with cinnamon, a natural fungicide, can help it dry and heal. \n\n<strong>Sunburn:</strong> This appears as a bleached, yellowish or whitish patch on stems that receive the most intense sun. While minor sunburn isn't fatal, it can stress the plant and create a weak spot for infections to enter. In extremely hot climates, providing some afternoon shade with a 30% shade cloth can be very beneficial. A healthy plant with good nutrition is also less susceptible to sunburn."] },
+            { heading: "Identifying and Controlling Common Pests", paragraphs: ["Dragon fruit is relatively tough, but a few pests can become a nuisance. \n\n- <strong>Ants:</strong> If you see ants marching up your plant, they are almost always a secondary issue. They are 'farming' other pests, like aphids or mealybugs, for their sweet secretions (honeydew). Get rid of the primary pest, and the ants will leave. \n\n- <strong>Mealybugs & Aphids:</strong> These small, sap-sucking insects often hide in the joints between stems. Look for white, cottony masses (mealybugs) or tiny green or black dots (aphids). For a minor infestation, a cotton swab dipped in rubbing alcohol will kill them on contact. For larger outbreaks, a spray of insecticidal soap or neem oil is effective. \n\n- <strong>Thrips and Mites:</strong> These are tiny pests that are hard to see, but you'll see their damage: scarred, scaly, or 'rusty' looking patches on the fruit and new growth. Regular preventative sprays with neem oil during the growing season is the best way to keep them at bay."], image: "https://placehold.co/800x600/FECACA/7F1D1D?text=Mealybugs+on+Stem+(800x600)" },
+            { heading: "The Best Defense: Natural and Preventative Methods", paragraphs: ["A healthy plant is a pest-resistant plant. The best way to deal with pests and diseases is to create an environment where they can't thrive. \n\n- <strong>Improve Air Circulation:</strong> Prune your plant annually to prevent a dense, tangled mess. Good airflow helps stems dry out quickly, making it harder for fungal spores to take hold. \n\n- <strong>Water at the Base:</strong> Avoid overhead watering. Splashing water from the soil onto the stems can transfer fungal spores directly to your plant. Water the soil, not the plant. \n\n- <strong>Regular Inspections:</strong> Spend a few minutes each week closely inspecting your plant, especially in the nooks and crannies. Catching a small pest problem early is much easier than dealing with a full-blown infestation. \n\n- <strong>Neem Oil is Your Friend:</strong> A regular, preventative spray of neem oil (every 2-4 weeks during the growing season) acts as a natural fungicide and insecticide, creating a first line of defense against most common issues."] },
+            { heading: "Dealing with Larger Pests (Birds and Rodents)", paragraphs: ["As your fruit begins to ripen, you may attract some larger, uninvited guests. Birds, squirrels, and even rats can be drawn to the sweet, colorful fruit. The most effective way to protect your harvest is with physical barriers. You can purchase 'organza bags' or 'fruit protection bags' online. These are small mesh bags that you can place over each individual fruit as it develops. This allows the fruit to ripen while keeping pests out. For larger-scale protection, you can drape bird netting over your entire trellis structure, securing it at the bottom so pests can't crawl under."] },
+        ],
+        blogLinks: [
+             {title: "5 Signs You're Overwatering Your Dragon Fruit", description: "Learn to recognize the subtle signs before it's too late.", path: "blog/5-signs-of-overwatering"},
+        ],
+        cta: { heading: "Keep Your Plants Healthy and Thriving!", subheading: "Our free calendar helps you track preventative care tasks.", buttonText: "Download Your Free Calendar" }
+};
