+++ content/guides/soil-guide.js
@@ -1,0 +1,19 @@
+export default {
+        title: "Best Soil for Dragon Fruit: The Ultimate DIY Mix Guide",
+        description: "Discover the perfect DIY soil mix for your dragon fruit. Learn about drainage, pH, and the best store-bought options for containers and in-ground planting.",
+        author: "Growing Dragon Fruit",
+        lastUpdated: "June 28, 2025",
+        crumbs: [{ label: "Soil Guide", path: "soil-guide" }],
+        toc: ["Why Soil is Critical", "The Three Pillars of Perfect Soil", "DIY Soil Mix Recipes", "Choosing Store-Bought Mixes", "Testing and Adjusting Soil pH", "Soil for Pots vs. In-Ground", "Amending Soil Seasonally"],
+        content: [
+          { heading: "Why Getting the Soil Right is Critical", paragraphs: ["To understand dragon fruit soil, you must understand the plant's nature. It is an <strong>epiphytic cactus</strong>, which means in its native jungle environment, it often grows on other trees, using its roots to anchor itself in bark crevices and pockets of decaying organic matter. This gives us our most important clue: it is not adapted to dense, wet, traditional garden soil. It hates having 'wet feet'. The right soil mix is the foundation of a healthy plant; it provides stability, prevents fatal root rot, allows roots to breathe, and delivers nutrients efficiently."] },
+          { heading: "The Three Pillars of Perfect Pitaya Soil", paragraphs: ["Every successful dragon fruit soil mix is built on three pillars: drainage, pH, and nutrients.", "<strong>1. Drainage:</strong> This is the most critical factor. The soil must be loose, gritty, and allow water to pass through freely. Without excellent drainage, water will stagnate around the roots, leading to rot. Key components for drainage include sand, perlite, pumice, and orchid bark.", "<strong>2. pH Level:</strong> Dragon fruit prefers a slightly acidic to neutral soil, with a target pH range between <strong>6.0 and 7.2</strong>. Most standard potting mixes and composts fall within this range, but it's wise to test your soil if you are having issues with nutrient uptake.", "<strong>3. Organic Nutrition:</strong> While it needs drainage, it's not a desert cactus. It thrives with a good amount of organic matter. Decomposed compost, worm castings, and aged manure provide a slow, steady release of the essential nutrients the plant needs for vigorous growth."] },
+          { heading: "Our Favorite DIY Soil Mix Recipes", paragraphs: ["Creating your own mix is easy, cost-effective, and gives you complete control. Here are two proven recipes:", "<strong>The Classic Mix (Balanced):</strong> \n- 40% Coarse Sand or Perlite (for drainage) \n- 40% Aged Compost or Worm Castings (for nutrients) \n- 20% Potting Soil or Coco Coir (for structure and moisture retention)", "<strong>The Pro-Potting Mix (For Containers):</strong> \n- 50% Orchid Bark Mix (contains bark, perlite, and charcoal) \n- 30% High-Quality Cactus/Succulent Mix \n- 20% Aged Compost", "As you can see in the image below, the ideal texture is loose and gritty, not dense and muddy."], image: "https://placehold.co/800x600/E9D5FF/3730A3?text=Gritty+Soil+Mix+(800x600)" },
+          { heading: "Choosing the Best Store-Bought Mix", paragraphs: ["If you prefer the convenience of a bagged mix, look for high-quality <strong>cactus and succulent soil</strong>. Read the ingredients list and avoid mixes that are very heavy on peat moss, as it can compact and retain too much moisture over time. Even with the best store-bought mix, we strongly recommend improving it. A simple and effective upgrade is to mix in one extra part of perlite or pumice for every two parts of the bagged soil. This small step can dramatically improve drainage and prevent future problems."] },
+          { heading: "Testing and Adjusting Your Soil's pH", paragraphs: ["If your plant is struggling with yellowing leaves despite proper fertilization, the issue could be pH. When soil pH is too high (alkaline) or too low (acidic), it 'locks up' nutrients, making them unavailable to the plant. You can buy a simple soil pH testing kit from any garden center. To lower pH (make it more acidic), you can amend the soil with elemental sulfur or use acid-loving plant fertilizers. To raise pH (make it more alkaline), you can add garden lime or wood ash. Always make small adjustments and re-test after a few weeks, as changing soil pH is a gradual process."] },
+          { heading: "Soil Considerations for Pots vs. In-Ground", paragraphs: ["For potted plants, drainage is paramount. Your container <strong>must</strong> have multiple large drainage holes. Do not add a layer of rocks at the bottom; this is a myth and actually hinders drainage. Use a light, airy soil mix like the 'Pro-Potting Mix' above. When planting in the ground, assess your native soil. If you have heavy clay, you must amend it heavily or, better yet, create a dedicated raised bed or mound for your dragon fruit. Dig a large hole (at least 2x2 feet) and backfill it entirely with your custom, well-draining dragon fruit soil mix."] },
+          { heading: "Amending and Revitalizing Your Soil Seasonally", paragraphs: ["The nutrients in a pot of soil are finite. Each year, it's a good practice to revitalize the soil, especially for hungry plants like dragon fruit. In the spring, you can gently top-dress your containers with a 1-2 inch layer of fresh compost or worm castings. This will provide a slow-release source of nutrients for the upcoming growing season. For in-ground plants, mulching with organic material serves a similar purpose, breaking down over time to enrich the soil."] },
+        ],
+        blogLinks: [],
+        cta: { heading: "Get Your Soil Right!", subheading: "Now that you have the foundation, learn how to feed your plant.", buttonText: "Learn About Fertilizing" }
+};
