+++ content/guides/trellis-guide.js
@@ -1,0 +1,24 @@
+export default {
+        title: "Dragon Fruit Trellis Designs: A DIY & Training Guide",
+        description: "Explore the best trellis designs for dragon fruit, including a DIY plan for the popular post and wheel system. Learn how to train your climbing cactus for maximum fruit production.",
+        author: "Growing Dragon Fruit",
+        lastUpdated: "June 28, 2025",
+        crumbs: [{ label: "Trellis Guide", path: "trellis-guide" }],
+        toc: ["Why a Trellis is Non-Negotiable", "Comparing Trellis Designs", "DIY Plan: The Post and Wheel Trellis", "Training Your Plant Up the Trellis", "Common Trellising Mistakes", "Alternative Trellis Ideas"],
+        content: [
+            { heading: "Why a Trellis is a Non-Negotiable Investment", paragraphs: ["A common beginner mistake is underestimating the size and weight of a mature dragon fruit plant. This is not a plant you can simply grow in a hanging basket. Dragon fruit is a vigorous climbing cactus that produces aerial roots to cling to surfaces, scaling its way towards sunlight. Without a strong support system, the plant will become a tangled, sprawling mess on the ground. This leads to a host of problems: the stems are more susceptible to rot and pests, air circulation is poor, and fruit production will be minimal or nonexistent.", "A proper trellis accomplishes several critical goals: it provides the necessary vertical support, it improves air circulation to prevent fungal diseases, and it allows the fruit-bearing branches to cascade downwards in an 'umbrella' shape. This cascading form is the ideal structure for maximizing sunlight exposure and making harvesting easy. Investing in a strong trellis from day one is investing in the long-term health and productivity of your plant."] , image: "https://placehold.co/800x600/DBEAFE/1E3A8A?text=Trellis+with+Branches+(800x600)"},
+            { heading: "Comparing Common Trellis Designs", paragraphs: ["While you can get creative, most successful trellis designs fall into a few categories:"], table: { headers: ["Design", "Pros", "Cons"], rows: [
+                        ["Post and Wheel", "Excellent air circulation; Easy to harvest from all sides.", "Requires a strong, deeply set post."],
+                        ["T-Bar", "Simple to construct from wood.", "Can get crowded if not pruned well."],
+                        ["Wall-Mounted / Fence", "Saves space; Can look very ornamental.", "Difficult to prune the back; Reduced air circulation."]
+                    ]}},
+            { heading: "DIY Plan: How to Build a Post and Wheel Trellis", paragraphs: ["This is a fantastic weekend project that will last for years. The video below gives a great overview of the process. \n\n<strong>Materials & Tools:</strong> \n- One 10-12 foot long, 6x6 inch pressure-treated wood or concrete post. \n- One bag of concrete mix. \n- A shovel and post-hole digger. \n- A level. \n- A sturdy top support (e.g., a repurposed bicycle rim with spokes removed, a welded rebar square, or a custom-built wooden cross). \n- Heavy-duty bolts or lag screws to attach the top support. \n\n<strong>Steps:</strong> \n1. **Dig the Hole:** Dig a hole at least 2-3 feet deep. The deeper, the better for stability. \n2. **Set the Post:** Place the post in the center of the hole. Use your level to ensure it is perfectly vertical. \n3. **Pour Concrete:** Mix the concrete according to the package directions and pour it into the hole, filling it completely. Brace the post so it remains level while the concrete cures (at least 24-48 hours). \n4. **Attach the Top:** Once the concrete is fully cured, securely bolt your chosen top support to the post. Your trellis is now ready for planting!"], video: { id: 's2j5h5Q6_pA', title: 'Building a DIY Dragon Fruit Trellis' } },
+            { heading: "Training Your Plant Up the Trellis", paragraphs: ["Patience is key during the training phase. After planting one to four cuttings at the base of your new trellis, select the strongest, most vigorous stem from each cutting to be the main 'trunk'. As this main stem grows, remove any other side shoots that sprout from its lower sections. This focuses all the plant's energy on reaching the top of the trellis. Use soft, flexible ties (like cloth strips, gardening velcro, or pantyhose) to loosely secure the stem to the post every foot or so. Do not tie it too tightly, as the stem will thicken over time. Once the main stem has grown about a foot higher than the top support, prune off the growing tip. This will stimulate the plant to branch out, creating the cascading umbrella of fruiting arms."] },
+            { heading: "Common Trellising Mistakes to Avoid", paragraphs: ["<strong>1. Building it Too Short:</strong> A trellis that is only 4-5 feet tall is not enough. Aim for at least 6-7 feet of height above the ground. \n<strong>2. Using Flimsy Materials:</strong> A mature dragon fruit plant laden with fruit is incredibly heavy. A thin wooden stake or a standard tomato cage will buckle and fail. Overbuild it from the start. \n<strong>3. Tying Stems Too Tightly:</strong> This can constrict and damage the stem as it grows, potentially killing the plant. The ties are just for guidance, not for strangulation."] },
+            { heading: "Alternative and Unconventional Trellis Ideas", paragraphs: ["Don't have space for a massive post? You can get creative. For smaller spaces, a very sturdy cattle panel or wire mesh arched between two posts can create a 'tunnel' trellis. Some growers successfully use the pillars of a patio or pergola as the main support, attaching a top structure to them. The key principles remain the same: provide a strong vertical support and a stable top platform for the branches to cascade over."] },
+        ],
+        blogLinks: [
+            {title: "A Step-by-Step Guide to Your First Pruning", description: "That first cut is the most important. Learn how to create the perfect shape for your plant.", path: "blog/pruning-for-shape"}
+        ],
+        cta: { heading: "Build the Perfect Support System!", subheading: "A good trellis is a long-term investment in your plant's health.", buttonText: "Explore Dragon Fruit Varieties" }
+};
