+++ content/guides/varieties.js
@@ -1,0 +1,19 @@
+export default {
+        title: "Dragon Fruit Varieties & Reviews: From American Beauty to Yellow Dragon",
+        description: "Explore popular dragon fruit varieties like American Beauty, Sugar Dragon, and Yellow Dragon. Learn about their flavor, pollination needs, and which is right for you.",
+        author: "Growing Dragon Fruit",
+        lastUpdated: "June 28, 2025",
+        crumbs: [{ label: "Varieties", path: "varieties" }],
+        toc: ["Pollination: The Most Important Factor", "American Beauty", "Sugar Dragon (S8)", "Yellow Dragon (Palora)", "Physical Graffiti", "Zamorano", "Where to Buy Authentic Cuttings"],
+        content: [
+            { heading: "Pollination: The First Thing to Consider", paragraphs: ["Before you fall in love with a variety's flavor or color, you must understand its pollination needs. This is the single most important factor for a home grower. Dragon fruit varieties are either <strong>self-fertile</strong> (can pollinate themselves and produce fruit with just one plant) or <strong>self-sterile</strong> (require pollen from a different dragon fruit variety to produce fruit).", "If you have limited space and only plan on growing one plant, <strong>you must choose a self-fertile variety.</strong> If you have space for multiple plants, you can explore the world of self-sterile varieties, which often have unique and complex flavors. However, you will need to plant at least two different varieties and be prepared to hand-pollinate them. In the guide below, we clearly state the pollination needs for each popular variety."] },
+            { heading: "American Beauty (Hylocereus guatemalensis)", paragraphs: ["<strong>Pollination:</strong> Self-Fertile \n\nAmerican Beauty is one of the most popular and reliable varieties for home growers, and for good reason. It's a vigorous, fast-growing plant that produces stunning magenta-fleshed fruit. The flavor is a step up from the common store-bought white-fleshed varieties, offering a refreshing sweetness with a hint of berry. It's known for being a consistent and heavy producer once established. Because it's self-fertile, it's a perfect choice for beginners who want a guaranteed harvest without the complexity of cross-pollination. Its bright pink skin and deep purple flesh make it a visual showstopper both on the vine and on the plate."], image: 'https://placehold.co/800x600/FBCFE8/831843?text=American+Beauty+(800x600)'},
+            { heading: "Sugar Dragon (S8)", paragraphs: ["<strong>Pollination:</strong> Self-Fertile \n\nAs the name implies, Sugar Dragon is renowned for its exceptional sweetness. Many enthusiasts consider it one of the best-tasting dragon fruits available. The fruit is typically smaller than American Beauty, but it packs an intense, rich, grape-like flavor with a high Brix (sugar content) level. The flesh is a deep, dark red. The plant itself is a moderately vigorous grower and is self-fertile, making it another excellent choice for those with a single plant. If your top priority is flavor intensity over fruit size, Sugar Dragon is an unbeatable variety."], image: 'https://placehold.co/800x600/F472B6/831843?text=Sugar+Dragon+(800x600)'},
+            { heading: "Yellow Dragon (Palora) (Hylocereus megalanthus)", paragraphs: ["<strong>Pollination:</strong> Self-Fertile \n\nThe 'queen' of sweet dragon fruits. The Yellow Dragon variety from Palora, Ecuador, is famous for its bright yellow skin and translucent white flesh. It is consistently rated as one of the sweetest and most delicious of all pitayas, with a flavor reminiscent of lychee, pear, and kiwi. The fruit contains larger, crunchier seeds than its pink-skinned cousins. While the plant's stems are thinner and more sensitive to intense sun than other varieties, its self-fertility and incredible flavor make it a highly sought-after plant for any serious hobbyist."], image: 'https://placehold.co/800x600/FEF08A/854D0E?text=Yellow+Dragon+(800x600)'},
+            { heading: "Physical Graffiti", paragraphs: ["<strong>Pollination:</strong> Self-Sterile (Requires a Pollinator) \n\nPhysical Graffiti is a perfect example of a variety that's worth the extra effort of cross-pollination. It produces beautiful fruit with a pink skin and a stunning purple-fuchsia flesh. The flavor is complex and excellent, often described as having berry and grape notes. To get fruit from Physical Graffiti, you must have another variety nearby to use as a pollen donor (American Beauty or Zamorano are good partners). You will need to hand-pollinate, transferring pollen from your donor plant to the Physical Graffiti flower. For gardeners with multiple plants, this variety is a must-have for its unique color and taste."], image: 'https://placehold.co/800x600/DDD6FE/4C1D95?text=Physical+Graffiti+(800x600)'},
+            { heading: "Zamorano", paragraphs: ["<strong>Pollination:</strong> Self-Fertile \n\nZamorano is a workhorse variety that is prized for its reliability and large, flavorful fruit. It produces impressive red-fleshed fruit that can often exceed one pound. The flavor is sweet and well-balanced. As a self-fertile variety, it's a low-maintenance choice for beginners, but it's also an excellent pollen donor for any self-sterile varieties you may want to grow, like Physical Graffiti. It's a vigorous grower and a great all-around choice for consistent production and great taste."] },
+            { heading: "Where to Buy Authentic Cuttings", paragraphs: ["The source of your cuttings matters. Buying from a reputable online nursery or a local grower who specializes in dragon fruit is crucial. Be wary of unidentified cuttings on auction sites, as you may end up with a low-quality, non-fruiting variety. Look for sellers who clearly label their varieties and can provide information about their pollination needs. A healthy cutting should be firm, green, and free of any soft spots or blemishes. Investing in a quality cutting from a good source is the best way to start your dragon fruit journey on the right foot."] },
+        ],
+        blogLinks: [],
+        cta: { heading: "Found Your Perfect Pitaya?", subheading: "Now learn exactly how to care for it with our complete guides.", buttonText: "Get Your Free Care Calendar" }
+};
