+++ content/index.js
@@ -1,0 +1,15 @@
+import howToGrow from './guides/how-to-grow.js';
+import soilGuide from './guides/soil-guide.js';
+import fertilizer from './guides/fertilizer.js';
+import trellisGuide from './guides/trellis-guide.js';
+import varieties from './guides/varieties.js';
+import pests from './guides/pests.js';
+
+export const pageContent = {
+  'how-to-grow': howToGrow,
+  'soil-guide': soilGuide,
+  'fertilizer': fertilizer,
+  'trellis-guide': trellisGuide,
+  'varieties': varieties,
+  'pests': pests,
+};
